moved bytecode to new runtime module
