+++ src/lib.rs
@@ -3,5 +3,6 @@ mod parser;
 mod symbol_map;
 mod ir;
 mod codegen;
+mod runtime;
 
 pub use driver::{execute, Config};
