+++ src/runtime/vm/mod.rs
@@ -1,0 +1,3 @@
+mod bytecode;
+
+pub use bytecode::{ByteCode, Func, Local, func_to_string};
